refactor: extract WebServer logic from FolioReaderContainer to EpubResourceServer
